This commit contains the changes to the desired models for the video embed.
