+++ app/forms/hyrax/thesis_form.rb
@@ -5,12 +5,14 @@
 module Hyrax
   class ThesisForm < Hyrax::Forms::WorkForm
     self.model_class = Thesis
+    self.terms += %i[
+      video_embed
+    ]
 
     self.terms -= DogBiscuits.config.base_properties
     self.terms += DogBiscuits.config.thesis_properties
 
     # Add any local properties here
-    self.terms += []
 
     self.required_fields = DogBiscuits.config.thesis_properties_required
