+++ app/models/generic_work.rb
@@ -1,6 +1,7 @@
 # frozen_string_literal: true
 
 class GenericWork < ActiveFedora::Base
+  include VideoEmbedViewer
   include ::Hyrax::WorkBehavior
   include DogBiscuits::Abstract
   include DogBiscuits::BibliographicCitation
@@ -20,20 +21,6 @@ class GenericWork < ActiveFedora::Base
   )
 
   validates :title, presence: { message: 'Your work must have a title.' }
-
-  # rubocop:disable Style/RegexpLiteral
-  validates :video_embed,
-            format: {
-              # regex matches only youtube & vimeo urls that are formatted as embed links.
-              with: /(http:\/\/|https:\/\/)(www\.)?(player\.vimeo\.com|youtube\.com\/embed)/,
-              message: "Error: must be a valid YouTube or Vimeo Embed URL."
-            },
-            if: :video_embed?
-  # rubocop:enable Style/RegexpLiteral
-
-  def video_embed?
-    video_embed.present?
-  end
 
   self.indexer = WorkIndexer
   self.human_readable_type = 'Work'
